+++ src/Symfony/Bundle/FrameworkBundle/CHANGELOG.md
@@ -13,6 +13,7 @@ CHANGELOG
  * Deprecate the `notifier.logger_notification_listener` service, use the `notifier.notification_logger_listener` service instead
  * Allow setting private services with the test container
  * Register alias for argument for workflow services with workflow name only
+ * Improved `bin/console debug:container --env-vars` and the `--env-var` option to display the number of occurrences of each environment variable in the container.
 
 6.2
 ---